api/requirements.txt
fastapi
uvicorn[standard]
pydantic
passlib[bcrypt]
python-jose